+++ src/clojure_lsp/config.clj
@@ -1,7 +1,10 @@
 (ns clojure-lsp.config
   (:require
     [clojure-lsp.shared :as shared]
-    [clojure.string :as string]))
+    [clojure.edn :as edn]
+    [clojure.java.io :as io]
+    [clojure.string :as string]
+    [taoensso.timbre :as log]))
 
 (def change-debounce-ms 300)
 
@@ -23,3 +26,48 @@
                                 :locals true
                                 :keywords true}
                      :canonical-paths true}}})
+
+(defn ^:private read-edn-file [^java.io.File file]
+  (try
+    (->> (slurp file)
+         (edn/read-string {:readers {'re re-pattern}})
+         shared/keywordize-first-depth)
+    (catch Exception e
+      (log/error "WARNING: error while reading" (.getCanonicalPath file) (format "(%s)" (.getMessage e))))))
+
+(defn get-property [p]
+  (System/getProperty p))
+
+(defn get-env [p]
+  (System/getenv p))
+
+(defn ^:private file-exists? [^java.io.File f]
+  (.exists f))
+
+(defn ^:private get-home-config-file []
+  (if-let [xdg-config-home (get-env "XDG_CONFIG_HOME")]
+    (io/file xdg-config-home ".lsp" "config.edn")
+    (io/file (get-property "user.home") ".lsp" "config.edn")))
+
+(defn ^:private resolve-home-config [^java.io.File home-dir-file]
+  (when (file-exists? home-dir-file)
+    (read-edn-file home-dir-file)))
+
+(defn ^:private resolve-project-configs [project-root ^java.io.File home-dir-file]
+  (loop [dir (io/file (shared/uri->filename project-root))
+         configs []]
+    (let [file (io/file dir ".lsp" "config.edn")
+          parent (.getParentFile dir)]
+      (if parent
+        (recur parent (cond-> configs
+                        (and (file-exists? file)
+                             (not (= (.getAbsolutePath home-dir-file) (.getAbsolutePath file))))
+                        (conj (read-edn-file file))))
+        configs))))
+
+(defn resolve-config [project-root]
+  (let [home-dir-file (get-home-config-file)]
+    (reduce shared/deep-merge
+            (merge {}
+                   (resolve-home-config home-dir-file))
+            (resolve-project-configs project-root home-dir-file))))