+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2021.02.26-13.58.48
+
 - Improve clojure-lsp config search checking always home dir considering XDG_CONFIG_HOME and project root up to system root (/) - Fixes #339.
 - Handle incremental text changes on `textDocument/didChange` notifications following LSP protocol, improving performance for huge files. 
 - Add clj-kondo version to --version and server-info.